It moves RequestLocationPermissionContentWidget to a shared folder
